backend/requirements.txt
Django==5.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
email-validator==2.1.0
sendgrid==6.10.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
djangorestframework==3.14.0
dj-rest-auth==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-oauth-toolkit==2.3.0
# Commenting out conflicting packages - install them separately
# awscli==1.33.5
# awsebcli==3.20.10 